chore: rename back to original project name & update information
