+++ .github/FUNDING.yml
@@ -1,3 +1,3 @@
 # These are supported funding model platforms
 
-# custom: ["https://afdian.net/@lgc2333/"]
+custom: ["https://afdian.net/@lgc2333/"]
