+++ .gitignore
@@ -2,8 +2,12 @@
 /.idea/
 /.run/
 /venv/
+/.venv/
 /testnb2/
 .pdm-python
 build/
 __pycache__
-/nonebot_plugin_picstatus_ng/res/picstatus-debug.html
+/nonebot_plugin_picstatus/res/picstatus-debug.html
+/.env
+/.env.*
+!/.env.example
