+++ nonebot_plugin_picstatus/collectors/__init__.py
