+++ nonebot_plugin_picstatus/collectors/bot.py
