+++ nonebot_plugin_picstatus/collectors/cpu.py
