+++ nonebot_plugin_picstatus/collectors/disk.py
