+++ nonebot_plugin_picstatus/misc_statistics.py
