+++ nonebot_plugin_picstatus/res/assets/default_avatar.webp
