+++ nonebot_plugin_picstatus/res/assets/default_bg.webp
