+++ nonebot_plugin_picstatus/res/js/globalThis.d.ts
