+++ nonebot_plugin_picstatus/res/js/init-global.js
