+++ nonebot_plugin_picstatus/res/js/load-plugin.js
