+++ nonebot_plugin_picstatus/templates/__init__.py
