+++ nonebot_plugin_picstatus/templates/default/__init__.py
