+++ nonebot_plugin_picstatus/templates/default/res/css/no-shadow.css
