+++ nonebot_plugin_picstatus/templates/default/res/css/theme-dark.css
