+++ nonebot_plugin_picstatus/templates/default/res/templates/index.html.jinja
