+++ nonebot_plugin_picstatus/templates/default/res/templates/macros.html.jinja
