+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "nonebot-plugin-picstatus-ng"
+name = "nonebot-plugin-picstatus"
 dynamic = ["version"]
 description = "A NoneBot2 plugin generates a picture which shows the status of current device"
 authors = [
@@ -27,14 +27,14 @@ readme = "README.md"
 license = { text = "MIT" }
 
 [project.urls]
-homepage = "https://github.com/wyf9/nonebot-plugin-picstatus-ng"
+homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-picstatus"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["nonebot_plugin_picstatus_ng"]
+packages = ["nonebot_plugin_picstatus"]
 
 [tool.hatch.version]
-path = "nonebot_plugin_picstatus_ng/__init__.py"
+path = "nonebot_plugin_picstatus/__init__.py"
